scripts/encode.fs: Initial attempts at file transfer in Forth
redforth is now complete enough that we can (and should) write its own
support scrirpts in Forth. In this case we need tooling to transfer files
from a host to a microcontroller. Currently we handle this by converting
the file for transfer into Forth source that we can execute on the
microcontroller.

For example:

    redforth include scripts/encode.fs \
             encode myscript.fs \
             bye > myscript.xfr
    picocom --send-cmd "ascii-xfr -sd" /dev/ttyACM0

From that point we can transfer the file to a Rasperry Pi Pico by
issuing MOUNT (to Forth) and C-x C-s (to picocom).